Move hexpansion manager LED activity into its own app
A couple of users have wanted to use these LEDs for something
that isn't hexpansion related, and this gives a hopefully
cleaner separation away from the hexpansion code.

I also considered putting this into the front pattern app rather
than being a separate process - I have no strong preference
either way.
